+++ packages/common/src/content/_internal/computeProps.ts
@@ -26,13 +26,9 @@ export const getExtentObject = (itemExtent: number[][]): IExtent => {
 
 export function deriveLocationFromItemExtent(itemExtent?: number[][]) {
   const location: IHubLocation = { type: "custom" };
-  const geometry: any = getExtentObject(itemExtent); // TODO: this needs to be fixed -tom
+  const geometry: any = getExtentObject(itemExtent);
   if (geometry) {
-    const convertedExtent = {
-      ...extentToPolygon(geometry),
-      type: "polygon",
-    } as unknown as Geometry;
-    location.geometries = [convertedExtent];
+    location.geometries = [geometry];
     location.spatialReference = geometry.spatialReference;
     location.extent = itemExtent;
   }