+++ packages/common/src/projects/_internal/computeProps.ts
@@ -3,10 +3,10 @@ import { UserSession } from "@esri/arcgis-rest-auth";
 import { getItemThumbnailUrl } from "../../resources";
 import { IHubProject } from "../../core";
 import { IModel } from "../../types";
-
 import { isDiscussable } from "../../discussions";
 import { processEntityFeatures } from "../../permissions/_internal/processEntityFeatures";
 import { ProjectDefaultFeatures } from "./ProjectBusinessRules";
+import { computeLinks } from "./computeLinks";
 
 /**
  * Given a model and a project, set various computed properties that can't be directly mapped
@@ -43,6 +43,8 @@ export function computeProps(
     model.data.settings?.features || {},
     ProjectDefaultFeatures
   );
+
+  project.links = computeLinks(model.item, requestOptions);
 
   // cast b/c this takes a partial but returns a full project
   return project as IHubProject;