+++ packages/common/src/projects/fetch.ts
@@ -2,7 +2,6 @@ import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { getItem, IItem } from "@esri/arcgis-rest-portal";
 
 import { getFamily } from "../content/get-family";
-import { getHubRelativeUrl } from "../content/_internal/internalContentUtils";
 import { IHubProject } from "../core/types";
 import { PropertyMapper } from "../core/_internal/PropertyMapper";
 import { getItemBySlug } from "../items/slugs";
@@ -18,12 +17,9 @@ import { computeProps } from "./_internal/computeProps";
 import { getPropertyMap } from "./_internal/getPropertyMap";
 import { unique } from "../util";
 import { getProp } from "../objects/get-prop";
-import { getItemThumbnailUrl } from "../resources/get-item-thumbnail-url";
-import { getItemHomeUrl } from "../urls/get-item-home-url";
-import { getItemIdentifier } from "../items";
-import { getRelativeWorkspaceUrl } from "../core/getRelativeWorkspaceUrl";
 import { listAssociations } from "../associations";
 import { getTypeByIdsQuery } from "../associations/internal/getTypeByIdsQuery";
+import { computeLinks } from "./_internal/computeLinks";
 
 /**
  * @private
@@ -127,18 +123,7 @@ export async function enrichProjectSearchResult(
 
   // Handle links
   // TODO: Link handling should be an enrichment
-  result.links.thumbnail = getItemThumbnailUrl(item, requestOptions);
-  result.links.self = getItemHomeUrl(result.id, requestOptions);
-  const identifier = getItemIdentifier(item);
-  result.links.siteRelative = getHubRelativeUrl(
-    result.type,
-    identifier,
-    item.typeKeywords
-  );
-  result.links.workspaceRelative = getRelativeWorkspaceUrl(
-    result.type,
-    identifier
-  );
+  result.links = computeLinks(item, requestOptions);
 
   return result;
 }